--- a/exporters/otlp/otlplog/otlploghttp/internal/transform/log.go
+++ b/exporters/otlp/otlplog/otlploghttp/internal/transform/log.go
@@ -25,11 +25,12 @@ func ResourceLogs(records []log.Record) []*lpb.ResourceLogs {
 		return nil
 	}
 
-	resMap := resourceLogsMap(records)
+	resMap := resourceLogsMapPool.Get().(map[attribute.Distinct]*lpb.ResourceLogs)
 	defer func() {
 		clear(resMap)
 		resourceLogsMapPool.Put(resMap)
 	}()
+	resourceLogsMap(&resMap, records)
 
 	out := make([]*lpb.ResourceLogs, 0, len(resMap))
 	for _, rl := range resMap {
@@ -44,11 +45,10 @@ var resourceLogsMapPool = sync.Pool{
 	},
 }
 
-func resourceLogsMap(records []log.Record) map[attribute.Distinct]*lpb.ResourceLogs {
-	out := resourceLogsMapPool.Get().(map[attribute.Distinct]*lpb.ResourceLogs)
+func resourceLogsMap(dst *map[attribute.Distinct]*lpb.ResourceLogs, records []log.Record) {
 	for _, r := range records {
 		res := r.Resource()
-		rl, ok := out[res.Equivalent()]
+		rl, ok := (*dst)[res.Equivalent()]
 		if !ok {
 			rl = new(lpb.ResourceLogs)
 			if res.Len() > 0 {
@@ -57,20 +57,20 @@ func resourceLogsMap(records []log.Record) map[attribute.Distinct]*lpb.ResourceL
 				}
 			}
 			rl.SchemaUrl = res.SchemaURL()
-			out[res.Equivalent()] = rl
+			(*dst)[res.Equivalent()] = rl
 		}
 		rl.ScopeLogs = ScopeLogs(records)
 	}
-	return out
 }
 
 // ScopeLogs returns a slice of OTLP ScopeLogs generated from recoreds.
 func ScopeLogs(records []log.Record) []*lpb.ScopeLogs {
-	scopeMap := scopeLogsMap(records)
+	scopeMap := scopeLogsMapPool.Get().(map[instrumentation.Scope]*lpb.ScopeLogs)
 	defer func() {
 		clear(scopeMap)
 		scopeLogsMapPool.Put(scopeMap)
 	}()
+	scopeLogsMap(&scopeMap, records)
 
 	out := make([]*lpb.ScopeLogs, 0, len(scopeMap))
 	for _, sl := range scopeMap {
@@ -85,11 +85,10 @@ var scopeLogsMapPool = sync.Pool{
 	},
 }
 
-func scopeLogsMap(records []log.Record) map[instrumentation.Scope]*lpb.ScopeLogs {
-	out := scopeLogsMapPool.Get().(map[instrumentation.Scope]*lpb.ScopeLogs)
+func scopeLogsMap(dst *map[instrumentation.Scope]*lpb.ScopeLogs, records []log.Record) {
 	for _, r := range records {
 		scope := r.InstrumentationScope()
-		sl, ok := out[scope]
+		sl, ok := (*dst)[scope]
 		if !ok {
 			sl = new(lpb.ScopeLogs)
 			var emptyScope instrumentation.Scope
@@ -100,11 +99,10 @@ func scopeLogsMap(records []log.Record) map[instrumentation.Scope]*lpb.ScopeLogs
 				}
 				sl.SchemaUrl = scope.SchemaURL
 			}
-			out[scope] = sl
+			(*dst)[scope] = sl
 		}
 		sl.LogRecords = append(sl.LogRecords, LogRecord(r))
 	}
-	return out
 }
 
 // LogRecord returns an OTLP LogRecord generated from record.