Part of #5196 

### Benchmarks

```console
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp/internal/transform
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
               │   old.txt   │               new.txt               │
               │   sec/op    │   sec/op     vs base                │
ResourceLogs-8   6.033µ ± 4%   5.249µ ± 8%  -13.00% (p=0.000 n=10)

               │    old.txt    │               new.txt                │
               │     B/op      │     B/op      vs base                │
ResourceLogs-8   10.602Ki ± 0%   8.299Ki ± 0%  -21.72% (p=0.000 n=10)

               │  old.txt   │              new.txt              │
               │ allocs/op  │ allocs/op   vs base               │
ResourceLogs-8   188.0 ± 0%   178.0 ± 0%  -5.32% (p=0.000 n=10)
```